Fix the ever elusive Minty Air error
ShulkerBoxBlockMixin.java line 26 was the issue. Resolved by removing the mixin as a whole as according to ( https://github.com/ReMod-Studios/Voidlands-Java/commit/9c95fb4545b869a17cba467bfa1142fe1269f6e6 )
the only reason this mixin existed was to fix an issue with the loottable(maybe?) of the custom shulker boxes.
I still live in fear of ShulkerBoxBlock.get(Dyecolor color), so keep in mind that that exists!